financial-calculator

+++ python_introduction/finance_calculator.py
@@ -1,9 +1,9 @@
-income = int(input("Enter your monthly income: "));
-expense = int(input("Enter your total monthly expenses: "));
-savings = income - expense;
+income = float(input("Enter your monthly income: "));
+expense = float(input("Enter your total monthly expenses: "));
+savings = float(income - expense);
 
 #Projected Savings = Monthly Savings * 12 + (Monthly Savings * 12 * 0.05)).
-projected_savings = savings * 12 + (savings * 12 * 0.05);
+projected_savings = float(savings * 12 + (savings * 12 * 0.05));
 
 print(f"Enter your monthly income: {income}");
 print(f"Enter your total monthly expenses: {expense}");
